+++ src/app/shared/services/is-dirty-guard.service.ts
@@ -5,14 +5,13 @@ import { marker as TEXT } from '@ngneat/transloco-keys-manager/marker';
 import * as _ from 'lodash';
 import { Observable } from 'rxjs';
 
-import { ModalComponent } from '~/app/shared/components/modal/modal.component';
 import { IsDirty } from '~/app/shared/models/is-dirty.interface';
-import { DialogService } from '~/app/shared/services/dialog.service';
+import { ModalDialogService } from '~/app/shared/services/modal-dialog.service';
 @Injectable({
   providedIn: 'root'
 })
 export class IsDirtyGuardService implements CanDeactivate<IsDirty> {
-  constructor(private dialogService: DialogService) {}
+  constructor(private modalDialogService: ModalDialogService) {}
 
   canDeactivate(
     component: IsDirty,
@@ -28,13 +27,11 @@ export class IsDirtyGuardService implements CanDeactivate<IsDirty> {
   }
 
   private showConfirmationDialog(): Observable<boolean> {
-    const modalRef = this.dialogService.open(ModalComponent, undefined, {
-      type: 'yesNo',
-      icon: 'danger',
-      message: TEXT(
+    const modalRef = this.modalDialogService.yesNoDanger(
+      TEXT(
         'You have made changes that have not been saved yet. Do you want to discard these changes?'
       )
-    });
+    );
     return modalRef.closed;
   }
 }